Mastering CentOS 7 Linux server: configure, manage, and secure a CentOS 7 Linux server to serve a variety of services provided in a sustainable computer's infrastructure - Alibi, Bhaskar Jyoti Roy
Key benefits
+  Learn how to efficiently set up and manage a Linux server using one of the best suited technologies for this purpose, CentOS 7
+  Personalize your Linux server and familiarize yourself with the latest tools and utilities setup provided by the new CentOS distribution
+  Follow a step-by-step tutorial through the configuration of the requested services with the capacity to personalize them as per your needs

Description
Most server infrastructures are equipped with at least one Linux server that provides many essential services, both for a user's demands and for the infrastructure itself. Setting up a sustainable Linux server is one of the most demanding tasks for a system administrator to perform. However, learning multiple, new technologies to meet all of their needs is time-consuming. CentOS 7 is the brand new version of the CentOS Linux system under the RPM (Red Hat) family. It is one of the most widely-used operating systems, being the choice of many organizations across the world. With the help of this book, you will explore the best practices and administration tools of CentOS 7 Linux server along with implementing some of the most common Linux services. We start by explaining the initial steps you need to carry out after installing CentOS 7 by briefly explaining the concepts related to users, groups, and right management, along with some basic system security measures. Next, you will be introduced to the most commonly used services and shown in detail how to implement and deploy them so they can be used by internal or external users. Soon enough, you will be shown how to monitor the server. We will then move on to master the virtualization and cloud computing techniques. Finally, the book wraps up by explaining configuration management and some security tweaks. All these topics and more are covered in this comprehensive guide, which briefly demonstrates the latest changes to all of the services and tools with the recent shift from CentOS 6 to CentOS 7.

Who is this book for?
If you are a Linux system administrator with an intermediate administration level, this is your opportunity to master the brand new distribution of CentOS. If you wish to possess a fully sustainable Linux server, with all its new tools and tweaks, that serves a variety of services to your users and customers, this book is ideal for you. It is your ticket to easily adapt to all the changes made in the latest shift.

What you will learn
+  Manage CentOS 7 users, groups, and root access privileges
+  Enhance the server's security through its firewall and prevent the most common attacks from penetrating or disabling the server
+  Explore and implement the common, useful services that a CentOS 7 server can provide
+  Monitor your server infrastructure for system or hardware issues
+  Create and configure a virtual machine using virtualization technologies
+  Implement a cloud computing solution on a single node system
+  Get an introduction to the configuration management tools and their usage
+  Discover the importance of the tools that provide remote connection, server service security, and system and process monitoring tools