Swift 3 Object-Oriented Programming: Implement object-oriented programming paradigms with Swift 3.0 and mix them with modern functional programming techniques to build powerful real-world applications , Second Edition - Gaston C. Hillar
Key benefits
+  Leverage the most efficient object-oriented design patterns in your Swift applications
+  Write robust, safer, and better code using the blueprints that generate objects
+  Build a platform with object-oriented code using real-world elements and represent them in your apps

Description
Swift has quickly become one of the most-liked languages and developers’ de-facto choice when building applications that target iOS and macOS. In the new version, the Swift team wants to take its adoption to the next level by making it available for new platforms and audiences. This book introduces the object-oriented paradigm and its implementation in the Swift 3 programming language to help you understand how real-world objects can become part of fundamental reusable elements in the code. This book is developed with XCode 8.x and covers all the enhancements included in Swift 3.0. In addition, we teach you to run most of the examples with the Swift REPL available on macOS and Linux, and with a Web-based Swift sandbox developed by IBM capable of running on any web browser, including Windows and mobile devices. You will organize data in blueprints that generate instances. You’ll work with examples so you understand how to encapsulate and hide data by working with properties and access control. Then, you’ll get to grips with complex scenarios where you use instances that belong to more than one blueprint. You’ll discover the power of contract programming and parametric polymorphism. You’ll combine generic code with inheritance and multiple inheritance. Later, you’ll see how to combine functional programming with object-oriented programming and find out how to refactor your existing code for easy maintenance.

Who is this book for?
This book is for iOS and macOS developers who want to get a detailed practical understanding of object-oriented programming with the latest version of Swift: 3.0.

What you will learn
+  Write high-quality and easy-to-maintain reusable object-oriented code to build applications for iOS, macOS, and Linux
+  Work with encapsulation, abstraction, and polymorphism using Swift 3.0
+  Work with classes, instances, properties, and methods in Swift 3.0
+  Take advantage of inheritance, specialization, and the possibility to overload or override members
+  Implement encapsulation, abstraction, and polymorphism
+  Explore functional programming techniques mixed with object-oriented code in Swift 3.0
+  Understand the differences between Swift 3.0, previous Swift versions, and Objective-C code